Revert "Records list: added symbols legend"
