+++ CHANGELOG.md
@@ -5,12 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.14] - 2025-06-20
+## [2.0.14] - TO-BE-RELEASED
 
 ### Added
 
 - Record editor: continue from last edited page;
-- Records list: added symbols legend;
 
 ### Fixed
