+++ package.json
@@ -11,9 +11,9 @@
   },
   "dependencies": {
     "@expo/webpack-config": "~19.0.1",
-    "@openforis/arena-core": "^1.0.29",
+    "@openforis/arena-core": "^1.0.28",
     "@react-native-async-storage/async-storage": "2.1.2",
-    "@react-native-clipboard/clipboard": "^1.16.2",
+    "@react-native-clipboard/clipboard": "^1.15.0",
     "@react-native-community/datetimepicker": "8.3.0",
     "@react-native-community/netinfo": "11.4.1",
     "@react-native-community/slider": "4.5.6",
@@ -72,7 +72,7 @@
     "react-native-version-number": "^0.3.6",
     "react-native-view-more-text": "^2.2.0",
     "react-native-web": "^0.20.0",
-    "react-native-zip-archive": "^7.0.2",
+    "react-native-zip-archive": "^7.0.1",
     "react-redux": "^9.2.0",
     "redux": "^5.0.1",
     "rn-swipe-button": "^3.0.1",
@@ -81,14 +81,14 @@
   "devDependencies": {
     "@babel/core": "^7.27.4",
     "@babel/eslint-parser": "^7.27.5",
-    "@eslint/compat": "^1.3.0",
+    "@eslint/compat": "^1.2.9",
     "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.29.0",
-    "@react-native/eslint-config": "^0.75.5",
+    "@eslint/js": "^9.28.0",
+    "@react-native/eslint-config": "^0.75.4",
     "babel-plugin-module-resolver": "^5.0.2",
     "babel-plugin-transform-remove-strict-mode": "^0.0.2",
     "dotenv": "^16.5.0",
-    "eslint": "^9.29.0",
+    "eslint": "^9.28.0",
     "eslint-plugin-ft-flow": "^3.0.11",
     "eslint-plugin-react": "^7.37.5",
     "metro-react-native-babel-preset": "^0.77.0",
