+++ src/localization/am/common.js
@@ -23,7 +23,6 @@ export default {
   goBack: "ተመለስ",
   info: "መረጃ",
   label: "መለያ",
-  legend: "አፈ ታሪክ",
   modifiedOn: "የተሻሻለው በ",
   name: "ስም",
   networkNotAvailable: "አውታረ መረብ የለም!",
