+++ src/localization/de/common.js
@@ -23,7 +23,6 @@ export default {
   goBack: "Zurück",
   info: "Info",
   label: "Bezeichnung",
-  legend: "Legende",
   modifiedOn: "Geändert am",
   name: "Name",
   networkNotAvailable: "Netzwerk nicht verfügbar!",
