+++ src/localization/en/common.js
@@ -23,7 +23,6 @@ export default {
   goBack: "Go back",
   info: "Info",
   label: "Label",
-  legend: "Legend",
   modifiedOn: "Modified on",
   name: "Name",
   networkNotAvailable: "Network not available!",
