+++ src/localization/fi/common.js
@@ -23,7 +23,6 @@ export default {
   goBack: "Palaa takaisin",
   info: "Info",
   label: "Nimiö",
-  legend: "Selite",
   modifiedOn: "Muokattu",
   name: "Nimi",
   networkNotAvailable: "Verkko ei ole käytettävissä!",
