+++ src/localization/id/common.js
@@ -23,7 +23,6 @@ export default {
   goBack: "Kembali",
   info: "Info",
   label: "Label",
-  legend: "Legenda",
   modifiedOn: "Diubah pada",
   name: "Nama",
   networkNotAvailable: "Jaringan tidak tersedia!",
