+++ src/localization/sv/common.js
@@ -23,7 +23,6 @@ export default {
   goBack: "Gå tillbaka",
   info: "Info",
   label: "Etikett",
-  legend: "Förklaring",
   modifiedOn: "Ändrad den",
   name: "Namn",
   networkNotAvailable: "Nätverk inte tillgängligt!",
