+++ src/screens/RecordsList/RecordsList.js
@@ -37,7 +37,6 @@ import { RecordsUtils } from "./RecordsUtils";
 import { RecordsListOptions } from "./RecordsListOptions";
 
 import styles from "./styles";
-import { RecordsListLegend } from "./RecordsListLegend";
 
 const { checkLoggedInUser } = RemoteConnectionUtils;
 
@@ -547,9 +546,6 @@ export const RecordsList = () => {
           </>
         )}
       </VView>
-
-      {syncStatusFetched && <RecordsListLegend />}
-
       {records.length > 0 && (
         <HView style={styles.bottomActionBar}>
           {newRecordButton}
